==> notebooks/01_download_and_validate_raw_data.ipynb <==
from pathlib import Path
import requests
import pandas as pd
import os

from taxi_demand_predictor.paths import DATA_DIR

def download_one_file_of_raw_data(year: int, month: int) -> Path:
    URL = f"https://d37ci6vzurychx.cloudfront.net/trip-data/yellow_tripdata_{year}-{month:02d}.parquet"
    response = requests.get(URL)

    if response.status_code == 200:
        path = f"{DATA_DIR}/raw/rides_{year}_{month:02d}.parquet"
        open(path, "wb").write(response.content)
        return path
    else:
        raise Exception(f"{URL} is not available")

def validate_raw_data(year: int, month: int) -> Path:
    rides = pd.read_parquet(f"{DATA_DIR}/raw/rides_{year}_{month:02d}.parquet")
    rides = rides[['tpep_pickup_datetime', 'PULocationID']]
    rides.rename(columns={
        'tpep_pickup_datetime': 'pickup_datetime',
        'PULocationID': 'pickup_location_id'
    }, inplace=True)
    rides = rides[rides.pickup_datetime >= '2022-01-01']
    rides = rides[rides.pickup_datetime < '2022-02-01']
    transformed_dir = DATA_DIR / "transformed" 
    if not transformed_dir.exists():
        transformed_dir.mkdir(parents=True, exist_ok=True)
    path = transformed_dir / f"validated_rides_{year}_{month:02d}.parquet"
    rides.to_parquet(path)
    return rides

download_one_file_of_raw_data(2022, 1)
# rides = clean_data(2022, 1)
