==> notebooks/02_transform_raw_data_into_ts.ipynb <==
from pathlib import Path
import pandas as pd

DATA_DIR = Path("..") / "data"

year = 2022
month = 1
rides = pd.read_parquet(f"{DATA_DIR}/transformed/validated_rides_{year}_{month:02d}.parquet")

rides.head(5)

# Add a column with the rounded hour
rides['pickup_hour'] = rides['pickup_datetime'].dt.floor('H')
rides

# Group by hours and count number of rows (per location id and pickup hour)
agg_rides = rides.groupby(['pickup_hour', 'pickup_location_id']).size().reset_index()
agg_rides.rename(columns={0: 'rides'}, inplace=True)
agg_rides

from tqdm import tqdm

def add_missing_slots(agg_rides: pd.DataFrame) -> pd.DataFrame:
    location_ids = agg_rides['pickup_location_id'].unique()
    full_range = pd.date_range(
        agg_rides['pickup_hour'].min(), agg_rides['pickup_hour'].max(), freq='H'
    )
    output = pd.DataFrame()
    for location_id in tqdm(location_ids):
        # keep only rides for this 'location_id'
        agg_rides_i = agg_rides.loc[agg_rides.pickup_location_id == location_id, ['pickup_hour', 'rides']]

        # Quick way to add dates with 0 in a Series
        agg_rides_i.set_index('pickup_hour', inplace=True)
        agg_rides_i.index = pd.DatetimeIndex(agg_rides_i.index)
        agg_rides_i = agg_rides_i.reindex(full_range, fill_value=0)

        agg_rides_i['pickup_location_id'] = location_id
        output = pd.concat([output, agg_rides_i])

    output = output.reset_index().rename(columns={'index': 'pickup_hour'})
    return output

agg_rides_all_slots = add_missing_slots(agg_rides)
agg_rides_all_slots

from typing import Optional, List
import plotly.express as px

def plot_rides(
        rides: pd.DataFrame,
        locations: Optional[List[int]] = None
):
    rides_to_plot = rides[rides.pickup_location_id.isin(locations) if locations else rides]

    fig = px.line(
        rides_to_plot,
        x="pickup_hour",
        y="rides",
        color="pickup_location_id",
        template='none',
    )

    fig.show()

plot_rides(agg_rides_all_slots, locations=[43])

agg_rides_all_slots.to_parquet(f"{DATA_DIR}/transformed/ts_rides_{year}_{month:02d}.parquet")
