==> notebooks/06_baseline_model.ipynb <==
## Create a baseline model without ML to obtain a baseline performance
## Metric error (MAE - Mean Absolute Error)

# Ways to improve model:

# 1. Increase training size
# 2. Add more features to training data
# 3. Try another algorithm (K-nearest, Random Forest, Logistic Regression, XGBoost, etc. etc.) - Tabular dataset best models: XGBoost, LightGBM, CatBoost 
# 4. Hyperparameter tuning (max_depth, learning_rate, n_estimators, gamma, etc. etc.)

import pandas as pd
from taxi_demand_predictor.paths import TRANSFORMED_DATA_DIR

df = pd.read_parquet(TRANSFORMED_DATA_DIR / "tabular_data.parquet")
df

from taxi_demand_predictor.data_split import train_test_split
from datetime import datetime

X_train, y_train, X_test, y_test = train_test_split(df, cutoff_date=datetime(2022,6, 1, 0, 0, 0), target_column_name='target_rides_next_hour')

print(f'{X_train.shape=}')
print(f'{y_train.shape=}')
print(f'{X_test.shape=}')
print(f'{y_test.shape=}')

import numpy as np

class BaselineModelPreviousHour:
    """
    Prediction = actual demand observed 1 hour ago
    """
    def fit(self, X_train: pd.DataFrame, y_train: pd.DataFrame):
        pass
    
    def predict(self, X_test: pd.DataFrame) -> np.array:
        return X_test[f'rides_previous_1_hour']

model = BaselineModelPreviousHour()
predictions = model.predict(X_test)
predictions

## How good are these predictions?
from sklearn.metrics import mean_absolute_error

test_mae = mean_absolute_error(y_test, predictions)
print(f'{test_mae=:.4f}')

# Predict the next hour using seasonal data
import pandas as pd
import numpy as np

class BaselineModelPreviousWeek:
    """
    Predict the next hour using the same hour of the previous week
    """
    def fit(self, X_train: pd.DataFrame, y_train: pd.Series):
        """
        Fit the model
        """
        pass

    def predict(self, X_test: pd.DataFrame) -> np.array:
        return X_test[f'rides_previous_{7*24}_hour']

model = BaselineModelPreviousWeek()
predictions = model.predict(X_test)

from sklearn.metrics import mean_absolute_error

test_mae = mean_absolute_error(y_test, predictions)
print(f'{test_mae=:.4f}')

# What about taking an average over the last month?

import pandas as pd
import numpy as np

class BaselineModelLast4Weeks:
    def fit(self, X_train: pd.DataFrame, y_train: pd.Series):
        pass
    
    def predict(self, X_test: pd.DataFrame) -> pd.Series:
        return 0.25 * (
            X_test[f'rides_previous_{7*24}_hour'] + \
            X_test[f'rides_previous_{2*7*24}_hour'] + \
            X_test[f'rides_previous_{3*7*24}_hour'] + \
            X_test[f'rides_previous_{4* 7*24}_hour']
        )

model = BaselineModelLast4Weeks()
predictions = model.predict(X_test)

test_mae = mean_absolute_error(y_test, predictions)
print(f'{test_mae=:.4f}')

# Final MAE = 2.9
