==> notebooks/07_xgboost_model.ipynb <==
%reload_ext autoreload
%autoreload 2


import pandas as pd
from taxi_demand_predictor.paths import TRANSFORMED_DATA_DIR

df = pd.read_parquet(TRANSFORMED_DATA_DIR / "tabular_data.parquet")
df

from datetime import datetime
from taxi_demand_predictor.data_split import train_test_split

# Split the data into train and test sets
X_train, y_train, X_test, y_test = train_test_split(df, cutoff_date=datetime(2022, 6, 1, 0, 0, 0), target_column_name='target_rides_next_hour')
print(f'{X_train.shape=}')
print(f'{y_train.shape=}')
print(f'{X_test.shape=}')
print(f'{y_test.shape=}')

import xgboost as xgb

# Only past rides for predictions
# So only get the columns that start with 'rides_'
# as they are the historical features
past_rides_columns = [c for c in X_train.columns if c.startswith('rides_')]
# Only take the columns that start with 'rides_'
X_train_only_numeric = X_train[past_rides_columns]

model = xgb.XGBRegressor()
# ONLY use historical data to train the model
model.fit(X_train_only_numeric, y_train)

X_test_only_numeric = X_test[past_rides_columns]
predictions = model.predict(X_test_only_numeric)
predictions

from sklearn.metrics import mean_absolute_error
test_mae = mean_absolute_error(y_test, predictions)
print(f'{test_mae=:.4f}')
